+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - APEL plugin: Fix bug in republish script when list of records is empty ([@dirksammel](https://github.com/dirksammel))
 - AUDITOR: Change the meta query field as an array ([@raghuvar-vijay](https://github.com/raghuvar-vijay))
 - Slurm collector: Speed up parsing of `sacct` output ([@rkleinem](https://github.com/rkleinem))
+- HTCondor collector: Allow for reading history from file ([@rfvc](https://github.com/rfvc))
 
 ### Removed